# API/openapi.yaml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Cryptour API
  license:
    name: SWE 2022
    
servers:
  - url: https://backend.cryptour.dullmer.de/
  - url: http://localhost:3030
tags:
  - name: Tours
    description: >-
      Tours are the core Dataset of the API. They describe a planned 
      hiking route and can be created, sold and bought through the planned
      crypto implementation 
  - name: Users
    description: >-
      A user can be a provider or consumer of a tour. 
paths:
  /tours:
    get:
      summary: List all tours
      operationId: listTours
      tags:
        - Tours
      parameters:
        - name: searchQuery
          in: query
          description: Search by title of the tour
          required: false
          schema:
            type: string
        - name: minDifficulty
          in: query
          description: Search by minimum tour difficulty
          required: false
          schema:
            type: integer
            format: int32
        - name: maxDifficulty
          in: query
          description: Search by maximum tour difficulty
          required: false
          schema:
            type: integer
            format: int32
        - name: minDistance
          in: query
          description: Search by minimum tour length
          required: false
          schema:
            type: integer
            format: int32
        - name: maxDistance
          in: query
          description: Search by maximum tour length
          required: false
          schema:
            type: integer
            format: int32
        - name: minTime
          in: query
          description: Search by minimum tour time 
          required: false
          schema:
            type: integer
            format: int32
        - name: maxTime
          in: query
          description: Search by maximum tour time 
          required: false
          schema:
            type: integer
            format: int32
        - name: location
          in: query
          description: Search by tour location
          required: false
          schema:
            type: string
      responses:
        200:
          description: An array of tours
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tours"
          headers:
            NumberOfResults:
              schema:
                type: number
              description: "How many Tours were found"
        400:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      summary: Create a tour
      operationId: createTour
      tags:
        - Tours
      security:
        - basicAuth: []
        - cookieAuth: []
      requestBody:
        description: A description of a Tour
        required: true
        content:
          application/json:
            schema: 
              $ref: "#/components/schemas/TourCreate"
      responses:
        201:
          description: Tour successfully created
        400:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /tours/{TID}:
    get:
      summary: Get a specific tour
      operationId: getTour
      tags:
        - Tours
      parameters:
        - name: TID
          in: path
          required: true
          description: The ID of the tour to retrieve
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Tour found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tour"
        400:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: Tour not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      summary: Delete a specific tour
      operationId: deleteTour
      tags:
        - Tours
      security:
        - basicAuth: []
        - cookieAuth: []
      parameters:
        - name: TID
          in: path
          required: true
          description: The ID of the tour to delete
          schema:
            type: integer
            format: int64
      responses:
        204:
          description: Tour deleted
        401:
          description: User does not own tour
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        403:
          description: Tour already bought by someone. Unable to delete.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: Tour not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /tours/{TID}/image:
    post:
      summary: Upload an image file
      operationId: uploadImage
      tags:
        - Tours
      security:
        - basicAuth: []
        - cookieAuth: []
      parameters:
        - name: TID
          in: path
          required: true
          description: The ID of the tour to delete
          schema:
            type: integer
            format: int64
      requestBody:
        description: The Image file
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        201:
          description: Image successfully uploaded
        400:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /tourc/{TID}/image/{IID}:
    get:
      summary: Get a tour image
      operationId: getTourImage
      tags:
        - Tours
      parameters:
        - name: TID
          in: path
          required: true
          description: The ID of the tour
          schema:
            type: integer
            format: int64
        - name: IID
          in: path
          required: true
          description: The ID of the image to retrieve
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Tour Image returned
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
        404:
          description: image not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        400:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /tours/{TID}/gpx:
    get:
      summary: Get a tour gpx file
      operationId: getTourGPX
      tags:
        - Tours
      security:
        - basicAuth: []
        - cookieAuth: []
      parameters:
        - name: TID
          in: path
          required: true
          description: The ID of the tour to retrieve
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Tour GPX returned
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        403:
          description: Tour not bought by current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        400:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        401:
          description: user has not bought tour
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: GPX not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      summary: Upload a gpx file
      operationId: uploadGPX
      tags:
        - Tours
      security:
        - basicAuth: []
      parameters:
        - name: TID
          in: path
          required: true
          description: The ID of the tour to delete
          schema:
            type: integer
            format: int64
      requestBody:
        description: The GPX file
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        201:
          description: GPX successfully uploaded
        400:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /tours/{TID}/rating:
    post:
      summary: Rate a tour
      operationId: rateTour
      security:
        - basicAuth: []
        - cookieAuth: []
      parameters:
        - name: TID
          in: path
          required: true
          description: The ID of the tour to retrieve
          schema:
            type: integer
            format: int64
      requestBody:
        description: The rating from 1 to 5
        required: true
        content:
          application/json:
            schema: 
              $ref: "#/components/schemas/Rating"
      tags:
        - Tours
      responses:
        200:
          description: Rating added or updated
        404:
          description: Tour not found
        406:
          description: Rating not between 1 and 5
        500:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      summary: Get a tour rating
      operationId: getTourRating
      tags:
        - Tours
      parameters:
        - name: TID
          in: path
          required: true
          description: The ID of the tour to retrieve
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Tour Rating returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rating"
          headers:
            NumberOfResults:
              schema:
                type: integer
                description: Number of submitted ratings
        400:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: tour not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /user:
    get:
      summary: Get info about myself
      operationId: getUser
      tags:
        - Users
      security:
        - basicAuth: []
        - cookieAuth: []
      responses:
        200:
          description: Successful request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        400:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      summary: Create a new User
      operationId: createUser
      tags:
        - Users
      requestBody:
        description: The user information
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegister"
      responses:
        201:
          description: User successfully created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        400:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      summary: Edit the information of an existing user
      operationId: patchUser
      tags:
        - Users
      security:
        - basicAuth: []
        - cookieAuth: []
      requestBody:
        description: The user information
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegister"
      responses:
        201:
          description: User successfully created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        400:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        500:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /user/createdTours:
    get:
      summary: Get my created tours
      operationId: getCreatedTours
      tags:
        - Users
      security:
        - basicAuth: []
        - cookieAuth: []
      responses:
        200:
          description: Tour found
          content:
            application/json:
              schema:
                type: array
                items:
                  type: number
        400:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /user/boughtTours:
    get:
      summary: Get my bought tours
      operationId: getBoughtTours
      tags:
        - Users
      security:
        - basicAuth: []
        - cookieAuth: []
      responses: 
        200:
          description: Bought tours
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tour"
        500:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /user/connectWallet:
    post:
      summary: Create a new User
      operationId: connectWallet
      tags:
        - Users
      security:
        - basicAuth: []
        - cookieAuth: []
      requestBody:
        description: The user wallet id on the Ethereum chain
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WalletConnect"
      responses:
        201:
          description: Wallet successfully connected
        401:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /user/login:
    post:
      summary: Log in user and receive session cookie
      tags:
        - Users
      security:
        - basicAuth: []
      responses:
        200:
          description: user successfully logged in
        401:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /user/logout:
    post:
      summary: destroy session cookie
      tags:
        - Users
      security:
        - cookieAuth: []
      responses:
        200:
          description: sessioncookie successfully destroyed
        401:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      
components:
  schemas:
    User:
      required:
        - firstname
        - surname
        - email
        - username
      properties:
        id:
          type: integer
          format: int64
        firstname:
          type: string
        surname:
          type: string
        username:
          type: string
        email:
          type: string
          format: email
        walletID:
          type: string
    UserRegister:
      required:
        - firstname
        - surname
        - email
        - username
        - password
      properties:
        firstname:
          type: string
        surname:
          type: string
        username:
          type: string
        password:
          type: string
        email:
          type: string
          format: email
        wallet-id:
          type: string
    Tour:
      properties:
        title:
          type: string
        difficulty:
          type: integer
          format: int32
        location:
          type: string
        distance:
          type: number
        duration:
          type: number
        description:
          type: string
        creatorID:
          type: string
        tourImages:
          type: array
          items:
            type: object
            properties:
              tiID:
                type: string
    TourCreate:
      properties:
        title:
          type: string
        difficulty:
          type: integer
          format: int32
        location:
          type: string
        distance:
          type: number
        duration:
          type: number
        description:
          type: string
    Tours:
      type: array
      items:
        $ref: "#/components/schemas/Tour"
    Rating:
      required:
        - rating
      properties:
        rating:
          type: integer
    WalletConnect:
      properties:
        walletAddress:
          type: string
    Error:
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
  responses:
    UnauthorizedError:
      description: Authentication information is missing or invalid
      headers:
        WWW_Authenticate:
          schema:
            type: string

  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    cookieAuth:
      type: apiKey
      in: cookie
      name: SessionID